training tasks/ELASTIC_SEARCH/Elastic Search with DRF/requirements.txt
Django==4.1.0
elasticsearch-dsl>=7.0.0
django-elasticsearch-dsl>=7.0.0
django-elasticsearch-dsl-drf>=0.22.0
djangorestframework>=3.14.0
python-dotenv==1.0.0
Pillow==9.4.0
django-debug-toolbar==3.8.1
django-elasticsearch-debug-toolbar==3.0.2
django-debug-toolbar-force==0.2